Split TiDBTypes.h
