+++ dbms/src/Storages/KVStore/tests/gtest_tikv_keyvalue.cpp
@@ -19,6 +19,7 @@
 #include <Storages/KVStore/TiKVHelpers/TiKVRecordFormat.h>
 #include <Storages/KVStore/tests/region_helper.h>
 #include <TestUtils/TiFlashTestBasic.h>
+#include <TiDB/Schema/TiDB.h>
 
 namespace DB
 {
