+++ dbms/src/TiDB/Decode/Datum.h
@@ -15,7 +15,8 @@
 #pragma once
 
 #include <Core/Field.h>
-#include <TiDB/Schema/TiDB.h>
+#include <TiDB/Schema/TiDBTypes.h>
+#include <TiDB/Schema/TiDB_fwd.h>
 
 #include <optional>
