+++ dbms/src/TiDB/Decode/DatumCodec.h
@@ -18,7 +18,8 @@
 #include <Core/Field.h>
 #include <IO/Buffer/WriteBuffer.h>
 #include <IO/Endian.h>
-#include <TiDB/Schema/TiDB.h>
+#include <TiDB/Schema/TiDBTypes.h>
+#include <TiDB/Schema/TiDB_fwd.h>
 
 /// Functions in this file are used for individual datum codec, i.e. UInt/Int64, Float64, String/Bytes, Decimal, Enum, Set, etc.
 /// The internal representation of a datum in TiFlash is Field.
